Move `ContentListViewModel` to `pillarbox-demo-shared`
